Fin de la creacion de la aplicacion

+++ src/hooks/useCalendarStore.js
@@ -1,5 +1,8 @@
 import { useDispatch, useSelector } from 'react-redux'
-import { onAddNewEvent, onDeleteEvent, onSetActiveEvent, onUpdateEvent } from '../store';
+import { onAddNewEvent, onDeleteEvent, onLoadEvents, onSetActiveEvent, onUpdateEvent } from '../store';
+import { calendarApi } from '../api';
+import { convertEventsToDateEvents } from '../helpers';
+import Swal from 'sweetalert2';
 
 
 export const useCalendarStore = () => {
@@ -7,6 +10,7 @@ export const useCalendarStore = () => {
 
     const dispatch = useDispatch();
     const { events, activeEvent } = useSelector( state => state.calendar )
+    const { user } = useSelector( state => state.auth )
 
     const setActiveEvent = ( calendarEvent )=> {
       dispatch( onSetActiveEvent( calendarEvent ) )
@@ -15,26 +19,65 @@ export const useCalendarStore = () => {
 
     const startSavingEvent = async( calendarEvent )=> {
 
-      // TODO: llegar al backend
-
-
-      // Todo bien
+      // TODO: Update Event 
       
-      if ( calendarEvent._id) {
-        // TODO: Actualizando
-          dispatch( onUpdateEvent( { ...calendarEvent } ) );
-      } else {
+      try {
+        if ( calendarEvent.id) {
+
+          // TODO: Actualizando
+          await calendarApi.put(`/events/${ calendarEvent.id }`, calendarEvent);
+          dispatch( onUpdateEvent( { ...calendarEvent, user } ) );
+          return; 
+        }
+     
         // Creando 
-        dispatch( onAddNewEvent( {...calendarEvent, _id: new Date().getTime() } ) );
+        const { data } = await calendarApi.post('/events', calendarEvent)
+        // console.log( {data} );
+        dispatch( onAddNewEvent( {...calendarEvent, id: data.evento.id, user } ) );
+
+      } catch (error) {
+        console.log( error );
+        Swal.fire('Error al guardar', error.response.data.msg, 'error' );
       }
+      
     };
 
 
-    const startDeletingEvent =  ()=> {
+    const startDeletingEvent =  async()=> {
 
 
-      //TODO: llegar al backend 
-      dispatch( onDeleteEvent() );
+      try {
+        if ( activeEvent.id) {
+
+          // TODO: Actualizando
+          await calendarApi.delete(`/events/${ activeEvent.id }`);
+          await calendarApi.get('/events')
+          dispatch( onDeleteEvent() )
+          return; 
+        }
+     
+
+      } catch (error) {
+        console.log( error );
+        Swal.fire('Error al Eliminar', error.response.data.msg, 'error' );
+      }
+
+    }
+
+
+
+
+    const startLoadingEvent = async() => {
+
+      try {
+        const { data } = await calendarApi.get('/events');
+        const event = convertEventsToDateEvents(data.eventos );
+        dispatch( onLoadEvents(event) );
+
+      } catch (error) {
+          console.log( 'Error cargar eventos' );
+          console.log( error );
+      }
     }
 
 
@@ -48,7 +91,8 @@ export const useCalendarStore = () => {
     //* Methods
     setActiveEvent,
     startSavingEvent,
-    startDeletingEvent
+    startDeletingEvent,
+    startLoadingEvent,
 
   }
 }
